sif: Factor mbus entity helpers and the lane-serving loop out of pci::serve
